Script: special processing for static array on SCMD_CALLOBJ instruction

+++ Engine/script/runtimescriptvalue.cpp
@@ -125,7 +125,7 @@ RuntimeScriptValue RuntimeScriptValue::ReadValue()
     }
     else if (this->Type == kScValDynamicObject)
     {
-        rval.SetLong(this->GetDynamicManager()->ReadInt16(this->Ptr, this->Value));
+        rval.SetLong(this->GetDynamicManager()->ReadInt32(this->Ptr, this->Value));
     }
     else
     {
